requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1

# LLM & LangChain (fix versions for Gemini v1)
langchain>=0.3.0
langchain-google-genai>=1.0.3
langchain-openai>=0.2.0
openai>=1.37.0
google-generativeai>=0.8.2

# File processing
pypdf>=4.2.0
python-docx>=1.1.0
pillow>=10.3.0
pytesseract>=0.3.10
openai-whisper>=20231117

# Background processing
celery>=5.3.6
redis>=5.0.0

# Cloud storage (MinIO/S3)
boto3>=1.34.0

# Data validation
pydantic>=2.7.0
email-validator>=2.1.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Testing
pytest>=8.2.0
requests>=2.31.0

# CrewAI agents
crewai[tools,google-genai]>=0.64.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# Payment Integration
stripe>=7.0.0
